requirements.txt
numpy==1.20.1
Flask==1.1.2
scikit-learn==0.24.1
pyyaml==5.1
xgboost==1.3.3
matplotlib==3.3.4
lightgbm==3.11
seaborn==0.11.0
joblib==1.0.1